requirements.txt
python-dotenv==1.0.0
pyyaml==6.0.1
fastapi == 0.115.4
uvicorn == 0.32.0
torch == 2.4.1
transformers == 4.44.2
pydantic == 2.8.2
pandas==1.5.3
numpy==1.23.5
pytest == 8.3.2
langchain == 0.3.21
langchain-openai == 0.3.9
langchain-huggingface == 0.1.2
langchain-core == 0.3.47
sentence_transformers == 3.4.0
faiss-cpu==1.7.4 
sqlite3
logging
